Changed to template classe to prepare for 64bits version
