+++ inc/M_extension.h
@@ -43,41 +43,298 @@ namespace riscv_tlm {
 /**
  * @brief Instruction decoding and fields access
  */
-    class M_extension : public extension_base<std::uint32_t> {
+    template<typename T>
+    class M_extension : public extension_base<T> {
     public:
 
         /**
          * @brief Constructor, same as base clase
          */
-        using extension_base::extension_base;
+        using extension_base<T>::extension_base;
 
         /**
          * @brief Decodes opcode of instruction
          * @return opcode of instruction
          */
-        op_M_Codes decode() const;
+        [[nodiscard]] op_M_Codes decode() const {
 
-        inline void dump() const override {
-            std::cout << std::hex << "0x" << m_instr << std::dec << std::endl;
+            switch (opcode()) {
+                case M_MUL:
+                    return OP_M_MUL;
+                    break;
+                case M_MULH:
+                    return OP_M_MULH;
+                    break;
+                case M_MULHSU:
+                    return OP_M_MULHSU;
+                    break;
+                case M_MULHU:
+                    return OP_M_MULHU;
+                    break;
+                case M_DIV:
+                    return OP_M_DIV;
+                    break;
+                case M_DIVU:
+                    return OP_M_DIVU;
+                    break;
+                case M_REM:
+                    return OP_M_REM;
+                    break;
+                case M_REMU:
+                    return OP_M_REMU;
+                    break;
+                    [[unlikely]] default:
+                    return OP_M_ERROR;
+                    break;
+            }
+
+            return OP_M_ERROR;
         }
 
-        bool Exec_M_MUL() const;
+        inline void dump() const override {
+            std::cout << std::hex << "0x" << this->m_instr << std::dec << std::endl;
+        }
 
-        bool Exec_M_MULH() const;
+        void Exec_M_MUL() const {
+            unsigned int rd, rs1, rs2;
+            std::int32_t multiplier, multiplicand;
+            std::int64_t result;
 
-        bool Exec_M_MULHSU() const;
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
 
-        bool Exec_M_MULHU() const;
+            multiplier = static_cast<std::int32_t>(extension_base<T>::regs->getValue(rs1));
+            multiplicand = static_cast<std::int32_t>(extension_base<T>::regs->getValue(rs2));
 
-        bool Exec_M_DIV() const;
+            result = static_cast<std::int64_t>(multiplier * multiplicand);
+            result = result & 0x00000000FFFFFFFF;
+            this->regs->setValue(rd, static_cast<std::int32_t>(result));
 
-        bool Exec_M_DIVU() const;
+            this->logger->debug("{} ns. PC: 0x{:x}. M.MUL: x{:d} * x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                extension_base<T>::regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
 
-        bool Exec_M_REM() const;
+        void Exec_M_MULH() const {
+            unsigned int rd, rs1, rs2;
+            std::int32_t multiplier, multiplicand;
+            std::int64_t result;
+            std::int32_t ret_value;
 
-        bool Exec_M_REMU() const;
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
 
-        bool process_instruction(Instruction &inst);
+            multiplier = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            multiplicand = static_cast<std::int32_t>(this->regs->getValue(rs2));
+
+            result = static_cast<std::int64_t>(multiplier) * static_cast<std::int64_t>(multiplicand);
+
+            ret_value = static_cast<std::int32_t>((result >> 32) & 0x00000000FFFFFFFF);
+            this->regs->setValue(rd, ret_value);
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.MULH: x{:d} * x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_MULHSU() const {
+            unsigned int rd, rs1, rs2;
+            std::int32_t multiplier;
+            std::uint32_t multiplicand;
+            std::int64_t result;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            multiplier = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            multiplicand = this->regs->getValue(rs2);
+
+            result = static_cast<std::int64_t>(multiplier * static_cast<std::uint64_t>(multiplicand));
+            result = (result >> 32) & 0x00000000FFFFFFFF;
+            this->regs->setValue(rd, static_cast<std::int32_t>(result));
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.MULHSU: x{:d} * x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_MULHU() const {
+            unsigned int rd, rs1, rs2;
+            std::uint32_t multiplier, multiplicand;
+            std::uint64_t result;
+            std::int32_t ret_value;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            multiplier = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            multiplicand = static_cast<std::int32_t>(this->regs->getValue(rs2));
+
+            result = static_cast<std::uint64_t>(multiplier) * static_cast<std::uint64_t>(multiplicand);
+            ret_value = static_cast<std::int32_t>((result >> 32) & 0x00000000FFFFFFFF);
+            this->regs->setValue(rd, ret_value);
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.MULHU: x{:d} * x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_DIV() const {
+            unsigned int rd, rs1, rs2;
+            std::int32_t divisor, dividend;
+            std::int64_t result;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            dividend = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            divisor = static_cast<std::int32_t>(this->regs->getValue(rs2));
+
+            if (divisor == 0) {
+                result = -1;
+            } else if ((divisor == -1) && (dividend == static_cast<std::int32_t>(0x80000000))) {
+                result = 0x0000000080000000;
+            } else {
+                result = dividend / divisor;
+                result = result & 0x00000000FFFFFFFF;
+            }
+
+            this->regs->setValue(rd, static_cast<std::int32_t>(result));
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.DIV: x{:d} / x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_DIVU() const {
+            unsigned int rd, rs1, rs2;
+            std::uint32_t divisor, dividend;
+            std::uint64_t result;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            dividend = this->regs->getValue(rs1);
+            divisor = this->regs->getValue(rs2);
+
+            if (divisor == 0) {
+                result = -1;
+            } else {
+                result = dividend / divisor;
+                result = result & 0x00000000FFFFFFFF;
+            }
+
+            this->regs->setValue(rd, static_cast<std::int32_t>(result));
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.DIVU: x{:d} / x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_REM() const {
+            unsigned int rd, rs1, rs2;
+            std::int32_t divisor, dividend;
+            std::int32_t result;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            dividend = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            divisor = static_cast<std::int32_t>(this->regs->getValue(rs2));
+
+            if (divisor == 0) {
+                result = dividend;
+            } else if ((divisor == -1) && (dividend == static_cast<std::int32_t>(0x80000000))) {
+                result = 0;
+            } else {
+                result = dividend % divisor;
+            }
+
+            this->regs->setValue(rd, result);
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.REM: x{:d} / x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        void Exec_M_REMU() const {
+            unsigned int rd, rs1, rs2;
+            std::uint32_t divisor, dividend;
+            std::uint32_t result;
+
+            rd = this->get_rd();
+            rs1 = this->get_rs1();
+            rs2 = this->get_rs2();
+
+            dividend = static_cast<std::int32_t>(this->regs->getValue(rs1));
+            divisor = static_cast<std::int32_t>(this->regs->getValue(rs2));
+
+            if (divisor == 0) {
+                result = dividend;
+            } else {
+                result = dividend % divisor;
+            }
+
+            this->regs->setValue(rd, static_cast<std::int32_t>(result));
+
+            this->logger->debug("{} ns. PC: 0x{:x}. M.REMU: x{:d} / x{:d} -> x{:d}({:d})",
+                                sc_core::sc_time_stamp().value(),
+                                this->regs->getPC(),
+                                rs1, rs2, rd, result);
+        }
+
+        bool process_instruction(Instruction &inst) {
+            this->setInstr(inst.getInstr());
+
+            switch (decode()) {
+                case OP_M_MUL:
+                    Exec_M_MUL();
+                    break;
+                case OP_M_MULH:
+                    Exec_M_MULH();
+                    break;
+                case OP_M_MULHSU:
+                    Exec_M_MULHSU();
+                    break;
+                case OP_M_MULHU:
+                    Exec_M_MULHU();
+                    break;
+                case OP_M_DIV:
+                    Exec_M_DIV();
+                    break;
+                case OP_M_DIVU:
+                    Exec_M_DIVU();
+                    break;
+                case OP_M_REM:
+                    Exec_M_REM();
+                    break;
+                case OP_M_REMU:
+                    Exec_M_REMU();
+                    break;
+                    [[unlikely]] default:
+                    std::cout << "M instruction not implemented yet" << "\n";
+                    inst.dump();
+                    //NOP(inst);
+                    sc_core::sc_stop();
+                    break;
+            }
+
+            return true;
+        }
 
     private:
 
@@ -85,8 +342,8 @@ namespace riscv_tlm {
          * @brief Access to opcode field
          * @return return opcode field
          */
-        inline std::int32_t opcode() const override {
-            return static_cast<std::int32_t>(m_instr.range(14, 12));
+        [[nodiscard]] inline std::uint32_t opcode() const override {
+            return static_cast<std::uint32_t>(this->m_instr.range(14, 12));
         }
 
     };
